Game now working correctly
