+++ client/public/game/game.js
@@ -1,17 +1,1 @@
-var config = {
-  type: Phaser.AUTO,
-  parent: 'phaser-example',
-  width: 800,
-  height: 600,
-  physics: {
-    default: 'arcade',
-    arcade: {
-      debug: false,
-      gravity: { y: 0 }
-    }
-  },
-  scene: 
-};
- 
-var game = new Phaser.Game(config);
- 
+
