+++ routes/sockets.js
@@ -17,8 +17,8 @@ var sockets = function(socket, io) {
     // create a new player and add it to our players object
     players[socket.id] = {
       rotation: 0,
-      x: 800 / 2,
-      y: 600 - 50,
+      x: Math.floor(Math.random() * 700) + 50,
+      y: Math.floor(Math.random() * 500) + 50,
       playerId: socket.id,
       team: (Math.floor(Math.random() * 2) == 0) ? 'red' : 'blue'
     };
